python/requirements.txt
cloudpickle==1.6.0
cycler==0.10.0
dask==2021.5.0
decorator==4.4.2
fsspec==2021.5.0
imageio==2.9.0
kiwisolver==1.3.1
locket==0.2.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.1.48
partd==1.2.0
picamera==1.13
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tifffile==2021.4.8
toolz==0.11.1
tqdm==4.60.0
